pyproject.toml
[tool.poetry]
name = "selenium-vs-playwright"
version = "0.1.0"
description = "A simple repo to compare Selenium & Playwright"
authors = ["wflazuardy <wflazuardy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.4.3"
playwright = "^1.25.2"
loguru = "^0.6.0"
webdriver-manager = "^3.8.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
